move supporting functions to `commit_message.py` and remove redundant `parse_bugs`
